Fix for FastSiPm fitting
We didnt have a way to set which function to fit, as a result every trace was being fit with the PMT Function. while this is fine for most of our detectors, its not for the SiPm Fast outputs (used @ ANL1471). This was causing the fits to fail (or just be bad) leading to a significant increase in the system ToF resolution.